Add the convergence+separation projection changes.
These will modify the one-eye projection matrix to take into account current separation and convergence settings.

This follows the example from the old stereo sample, to use the settings and generate the proper sized results.  Also sets the XM_NO_INTRINSICS so that we have access to the _41 and _31 parameters.  This is a sample program, and performance is not as important as clarity.

This works to draw the cube fully in 3D, both eyes working, with proper depth, and proper convergence.  The hot keys for depth and convergence also work, as it's updated every frame.

Fully functional.