Simplifying the sample to bare minimums.
The goal of this branch is to demonstrate the absolute bare minimum required for Direct Mode use, and avoid any spurious or confusing details.

This removes the color changing of the cube, and the texture mapped onto the surface.

The pixel shader just returns a grey pixel now for the color.