Restore the CreateDepthStencilView for stereo.
This requires that the DepthStencil texture be doubled in width as well, in order for Direct Stereo to work.  We expect that to be part of a game's job, so 3Dmigoto will handle the doubling of the width.

This is tested and working, for cases where the bind flags = D3D11_BIND_DEPTH_STENCIL.