Add Nektra In-Proc libraries for testing debug.
Found a bug with x64 In-Proc hook, this branch is for testing this particular bug.

Add code to hook the nvapi_QueryInterface call out of nvapi64.dll, which demonstrates the bug.

Includes the broken version here of NtkHookLib64_Debug.lib. We are only looking at the x64 Debug version here.

This version demonstrates the crash at HookedQueryInterface.