Move DX11 init to first, and use DXGI_PRESENT_TEST.
This is an interesting result, with the display showing the DX9 data, as an alternating colors for each eye.  Just using DX9->Clear for the moment, for each eye, but each eye definitely gets a different color.

Also this still calls DX11->Present, but uses the DXGI_PRESENT_TEST flag, which doesn't actually present, but still follows all the DX11 semantics.

So... an interesting test case.  Might work.