Switch to DX9Ex creation and use.
This sets up using the DX9Ex objects instead, as they allow more options and should presumably be more compatible and up to date.

Cannot use the D3DSWAPEFFECT_FLIPEX mode however, as two devices cannot target the same window with that mode.

If we use D3DSWAPEFFECT_FLIP for the DX9 output, and set the DX11 present to DXGI_PRESENT_TEST, then we get stereo output from the DX9 device without errors.  A little weird, but maybe this works.