Set DX11 swapchain to use a child window. Working.
This seems like a good path.  It sets up the app to use as the primary window output the DX9 swapchain, and then create an artificial child window of that, for the DX11 swapchain to target.

This seems to work as desired.  No errors are reported, and the output is running in stereo with different colors in each eye, when viewed with 3D Vision glasses.

This seems better than using The DXGI_PRESENT_TEST for DX11, as it's a fully supported approach.