Move DX11 init to subroutine, for fullscreen handling.
When switching to fullscreen, we must call ResizeBuffers, which requires everything to be Released.  This sets up creation and init to different routines, so that init can be done at every alt-return.

Sort of works. Goes into full screen, but fails on exit, because dx11 is mad about the child window.