Tweak dx11 params to handle fullscreen properly.
This now runs completely, no errors.  Can move into or out of fullscreen mode without problems.  Starts as windowed.

ResizeBuffers needs full screen size, so that it matches the depth stencil size.

If we set DXGI_SWAP_EFFECT_DISCARD, then it all works. If we use DXGI_SWAP_EFFECT_FLIP_SEQUENTIAL we get an error at Present that a ResizeBuffers has not be done, but it definitely has.  If the window is not marked as a child, then we can DXGI_SWAP_EFFECT_FLIP_SEQUENTIAL and get no error.  Some interaction between child window and flip modes.