Remove dx11 swapchain fullscreen.
Because we are using DX9 as the primary window, there is no reason to switch dx11 into full screen mode.  It's an invisible child window, and as such it doesn't need fullscreen treatment.

This was causing the main window to be sent to the back, as the dx11 was taking control.

With this change it now switches fullscreen and windowed without errors.