Add some microsecond timer debug output.
These sample and report each eye and full frame timing, to be sure it's working as expected to sync to vertical blank.

Getting 8.33ms per eye, which is right. If 3D Vision is enabled, it drops to 16.66 instead though, which suggests that 3D Vision forces everyone else to see 60Hz.