Add the ShutterGlasses class from 3DVisionActivator.
https://github.com/FlintEastwood/3DVisionActivator

This is from a project created by cms, and shared on the mtbs3d forums that directly runs the NVidia 3D emitter pyramid via USB commands.

Since it's already a class and already running, we just make a copy of the code here.

The full project is very similar to this 3D-Vision-Direct, but has all sorts of weird dependencies that don't build.