Switch to Multi-Byte from UTF.
The shutter glasses object is built around ASCII to pipe to USB, and it's uphill both ways to make it UTF.

Just setting this variant to Multi-Byte like the original code, as it's just testing for now. Switch our string uses from UTF to ASCII.

Add the setupapi.lib to the linker so that we get the USB functions.