And... it works!
Add the NVidiaShutterGlasses object to the main app, and call the eyeswap/sync after each eye has been Present.

Works! Using the MonitorInfo as specified, this is fully in sync as well. Emitter lights up bright green while in use.

Since the Present(1,0) blocks until the vertical blank, the toggleEyes happens immediately after the vertical blank.

Can get an eye-swap if a frame is dropped for any reason, like dragging the window or alt-tab.

But hell yeah, it works!