Good checkpoint, best no-eye-swap so far.
This will still swap eyes, but seems quite a lot better when doing alt-tab. It will flicker glasses, but most often it will restart with left eye in sync.

This is just starting with setLeftEye at Init time, then falling into the Render loop where it calls setRightEye after both of the Present calls. So only one setRightEye for both eyes drawn.

This seems to work because of the emitter design, where it is free running on it's own. The setRightEye just seems to synchronize, and does not actually set either eye.