Remove emitter Read operation.
Since that can crash the system, and we don't actually give a rip about the wheel or button on the front, let's just not ever read from the emitter.

Presumably the clear() operation will dump any pending send buffer, and in any case it seems to not care that we never read.

clear() does not provide any improvement in sync. Dragging window will flip eyes, alt-tab restores them. Does not appear to be any way to avoid eye flips here.