Add a Wake setup for USB, to fix BSOD.
The emitter can be in sleep mode by Win10 nowadays, and if we address it normally we'll get a BSOD. It's not quite clear, but seems like it does not wake fast enough.

This adds an extra pipe operation, as I noticed that it will wake the emitter upon simply opening the USB pipe, even though that pipe is non-functional.

This just has a wake pipe that is open, wait for 1 second, then close that pipe. This wakes the device and avoids the crash.