Add enable_LightBoost_NVIDIA routine.
And disable_LightBoost_NVIDIA when exiting in destructor.

This is trying to be monitor safe, and currently only works for PG278QR when set as primary monitor and running at 120Hz.

IFF that is true after fetching monitor info, we set the VTotal pixels in NV_TIMING to +5, which tells the monitor to enable LightBoost.

This works! LightBoost is on, and the green led in lower right is lit. OSD says it's in 3D mode, not ULMB.

Timing is now off for MonitorTimings, phasing in and out, presumably because of those five extra lines.