Finally fix timing problem with LightBoost on.
OK, that took some real effort. The problem was that by bumping the back porch by 5 vertical lines, we altered the frame timing.

The answer is to *also* bump the pixelclock to match the new timing of pixels, while leaving the refresh rate itself at 120Hz.

Doing this now works completely. LightBoost comes on while app is running, glasses startup and sync, image is clear. Nice!