Add F2 to swap eyes.
Tweak to use toggleEyes instead of specific setting, and add F2 to swap eyes in case of bad swaps.

This works, surprisingly. F2 will alter the current state, and the glasses will then toggle as expected.

If I keep deliberate setLeftEye, and setRightEye for each Present then the F2 swap no longer works. So it clearly is paying some attention.