Remove parameter 'offset' from toggleEyes.
This doesn't really make any sense to have a parameter, it's always the same number, which is also not an offset. It's a chunk of the actual command bytes to the emitter.

This commit changes the routines to just be called, no parameters, and puts the 'offset' bytes of 0xffff0000 directly into the sequence array.

Tested as running correctly, nothing mysterious about these bytes, but we do have to send the exact right count in the correct order.