Convert project to Unicode Character Set.
The geo-11 code is set up with Unicode already, and we don't want to change that.

This commit changes the CharacterSet from Multi-Byte to Unicode, and converts everything to use wstring like we do in 3Dmigoto.

A little scary changing the USB comms- but it worked fine to convert to wstring and use the automatic W versions of the calls.

Tested as working correctly.