Add EDID checking for LightBoost.
This adds a whole gnarly NVidia sequence to fetch the EDID from the connected monitor. It's a raw byte array with a bunch of encoding.

If the EDID that we parse from the live monitor matches the record in the MonitorTimings.ini, we consider that OK and enable LightBoost via bumping back porch timing by +5.

Tested as working correctly on ASUS PG278QR, which returns the AUS_27B1 EDID.