A big raft of experimental fixes for Present. Mostly useless.
Lots of changes suggested by ChatGPT to experiment with different Present variants.

Upshot is that it Present as currently implemented is useless, and doing sync via Present is a fools errand.

All the Present(0,0) does is immediately show the buffer, whether tearing is disabled or not. I conclude that Present variants are all useless, and the only variant that matters at at all is Present(1,0).

I'll likely remove some of this debugging/experimental stuff, but it still runs now, so it's not broken.