Add an output swapchain and device for the refresh thread.
Since this cannot be the original drawing thread, we need a second swap chain to be able to Present.

This sets up a couple of R/L eye textures, and sends them to the backbuffer in the refresh thread.  The main drawing thread is running, but not showing right now.

This is functional, so it's a good checkpoint.