Create shared surfaces with D3D11_RESOURCE_MISC_SHARED.
Working!

Changes the destination for cube drawing to be two Texture2D resources, to emulate the geo-11 sharedRTV.  Those are created with the D3D11_RESOURCE_MISC_SHARED flag set so that we can then OpenSharedResource in the RefreshThread and get access to the surface on the different ID3D11Device used to match sync on the monitor.

Seems to be working very well. No errors from debug layer, glasses sync properly.

One flaw- still saw a couple of eye swaps.