Put drawing loop back like it was originally, drawing on main thread.
In the WinMain loop, we call render_frame, which will draw into the two shared Texture2D surfaces. This is independent of the output view.

Also enables lag sleep test- where we simulate slow drawing by making the draw loop sleep for 33ms.

This works. Drawing and syncing to eyes is still happening, even as the cube is rotating slowly and jerkily.

Still get eye swaps though, which should not be possible.