Working checkpoint with mutex.
Not positive we need the mutex, but when free running with tearing, we otherwise get glitchy graphics.

This is running with a std::mutex to block drawing or reading from the shared surfaces if they are in use by the other thread. Seems to work well.

Still get eye swaps, and the drawing itself never swaps, so it must be that the glasses time out or get confused. Even with a raised thread priority we get eye swaps.