Handle exclusive fullscreen state.
This seems like something we probably need, although this is fighting an uphill battle because of the extra thread for output.

If the DWM is genuinely interfering with the drawing though, we can put it on hold via FullScreenState, so it's worth the experiments.

This commit sets up and handles the fullscreen transition properly now, and includes some gotcha fixes from the threading, like requiring a SetFullScreenState(false,0) on the main thread, otherwise it hangs.

Fully running. Can transition back and forth.